+++ tests/testthat/test-clear_cache.R
@@ -1,18 +1,22 @@
-# Tests depend on PIPAPI_DATA_ROOT_FOLDER. Skip if not found.
-skip_if(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER") == "")
+# Tests depend on PIPAPI_DATA_ROOT_FOLDER_LOCAL. Skip if not found.
+skip_if(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER_LOCAL") == "")
+#
 
-lkups <- create_versioned_lkups(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER"))
-lkups <- lkups$versions_paths[[lkups$latest_release]]
+lkups <- create_versioned_lkups(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER_LOCAL"))
+lkup <- lkups$versions_paths[[lkups$latest_release]]
 
-# Tests depend on a specific temp cache directory. Skip if not found.
-cd <- readRDS('../testdata/cd-ex.RDS')
+# lkup_path <- test_path("testdata", "lkup.rds")
+# lkup      <- readRDS(lkup_path)
+
+cd        <- readRDS(test_path("testdata", "cd-ex.RDS"))
+
 skip_if_not(dir.exists(cd$info()$dir),
             message = sprintf("The cache directory '%s' doesn't exist", cd$info()$dir))
 
 test_that("clear_cache() is working as expected", {
 
   # Send request to make sure that at least one item is in the cache
-  pip("AGO", 2008, lkup = lkups)
+  pip("AGO", 2008, lkup = lkup)
 
   # items found
   res <- clear_cache(cd)
@@ -30,3 +34,5 @@ test_that("clear_cache() is working as expected", {
   expect_equal(res$msg,  'Cache directory not found.')
 
 })
+
+