+++ NEWS.md
@@ -1,3 +1,16 @@
+# pipapi 1.2.0
+
+## New features
+- [Add new endpoint to support country profile downloads](https://github.com/PIP-Technical-Team/pipapi/pull/283)
+- [Add separate arguments to select release and ppp version](https://github.com/PIP-Technical-Team/pipapi/issues/265)
+- [Allow direct selection of non-official WB regions](https://github.com/PIP-Technical-Team/pipapi/pull/287)
+- [Add new /citation endpoint](https://github.com/PIP-Technical-Team/pipapi/pull/302)
+ 
+## Enhancements
+- [Disable popshare option for pip_grp and aggregate distributions](https://github.com/PIP-Technical-Team/pipapi/pull/274)
+- [Allow /aux endpoint to return data in long format](https://github.com/PIP-Technical-Team/pipapi/issues/272)
+- [Make default poverty line dynamic in API](https://github.com/PIP-Technical-Team/pipapi/pull/310)
+
 # pipapi 1.1.0
 
 ## Enhancements
@@ -49,7 +62,7 @@ interpolated years](https://github.com/PIP-Technical-Team/pipapi/issues/182)
 - Make sure `ui_cp_poverty_charts()` only returns non-national observations when a country has no surveys with national coverage
 - Fix a bug in the application of censoring within `pip_grp()`
 - Add a specific empty repsonse for `pip_grp()` to ensure that the response is consistent when no data is available
-- Fix a bug in the selection of most recent value (`year="mrv"`) when `country="all"`
+- Fix a bug in the selection of most recent value (`year="MRV"`) when `country="ALL"`
 
 # pipapi 0.0.1